% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rnetcarto.R
\name{bipartmod}
\alias{bipartmod}
\title{Computes modularity and modularity roles from a bipartite network.}
\usage{
bipartmod(web)
}
\arguments{
\item{web}{a matrix representing the interactions observed between
two groups of species (respectively rows and columns).}
}
\value{
A list. The first element is a dataframe with the name,
module, z-score, and participation coefficient for each row of the
input matrix. The second element is the modularity of this
partition.
}
\description{
Computes modularity and modularity roles from a bipartite network.
}

