% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rnetcarto.R
\name{bipartmod}
\alias{bipartmod}
\title{Computes modularity and modularity roles from a bipartite network.}
\usage{
bipartmod(web, seed = as.integer(floor(runif(1, 1, 100000001))),
  iterfac = 1, coolingfac = 0.995, degree_based = FALSE,
  weighted = TRUE)
}
\arguments{
\item{web}{a matrix representing the interactions observed between
two groups of species (respectively rows and columns).}

\item{seed}{Seed for the random number generator: Must be a
positive integer.}

\item{iterfac}{At each temperature of the simulated annealing
(SA), the program performs fN^2 individual-node updates (involving the
movement of a single node from one module to another) and fN
collective updates (involving the merging of two modules and the split
of a module). The number "f" is the iteration factor.}

\item{coolingfac}{Temperature cooling factor.}

\item{degree_based}{If TRUE, use the degree based roles metrics otherwise
use the strength based metric.}

\item{weighted}{If TRUE, use the weighted modularity definition.}
}
\value{
A list. The first element is a dataframe with the name,
module, z-score, and participation coefficient for each row of the
input matrix. The second element is the modularity of this
partition.
}
\description{
Computes modularity and modularity roles from a bipartite network.
}

